restructuring pages and layouts
