+++ .gitignore
@@ -1,5 +1,6 @@
 # Logs
 logs
+
 *.log
 npm-debug.log*
 yarn-debug.log*
@@ -57,3 +58,6 @@ typings/
 # dotenv environment variables file
 .env
 .idea
+
+MacOS
+.DS_Store
