+++ src/postinit.js
@@ -11,9 +11,9 @@ import {
 
 import { getProvider, createStore } from './redux';
 
-import Page from './layouts/Page';
-import IndexPage from './layouts/IndexPage';
-import FourOFourPage from './layouts/404Page';
+import Page from './pages/Page';
+import IndexPage from './pages/IndexPage';
+import FourOFourPage from './pages/404Page';
 import Routes from './routes';
 
 export default () => {
